ct/gc: move calculate pressure logic into GC
This commit moves the function `ctmap.CalculateCTMapPressure` into
the GC package and makes it a method of the struct `GC`. This way
we keep the GC related logic in that package.
